requirements.txt
opencv-python
pyyaml
tqdm
ultralytics
transformers
torch
pillow
requests
python-dotenv